split endpoints

+++ src/app/project/[project]/ManageData.tsx
@@ -1,57 +1,75 @@
-import { DataPage } from '@/types';
-import { useState } from 'react';
-import { FaArrowLeft, FaArrowRight } from 'react-icons/fa';
+import { DataPage } from "@/types";
+import { useState } from "react";
+import { FaArrowLeft, FaArrowRight } from "react-icons/fa";
 
 export default function ManageData({
   dataPage,
   deleteData,
-  setStatus
+  setStatus,
 }: {
   dataPage: DataPage;
-  deleteData: () => void;
+  deleteData: any;
   setStatus: (status: { loading: string; error: string }) => void;
 }) {
   const [canDelete, setCanDelete] = useState(false);
 
   function onDelete() {
-    setStatus({ loading: 'Deleting', error: '' });
+    if (!deleteData) return;
+    setStatus({ loading: "Deleting", error: "" });
     deleteData({})
       .then(() => setCanDelete(false))
       .finally(() => {
-        dataPage.resetData();
-        setStatus({ loading: '', error: '' });
+        dataPage.reset();
+        setStatus({ loading: "", error: "" });
       });
   }
 
-  if (!dataPage.data?.length) return null;
+  function changePage(e: MouseEvent, direction: "next" | "prev") {
+    e.preventDefault();
+    const fun = direction === "next" ? dataPage.nextPage : dataPage.prevPage;
+    fun?.();
+  }
+
+  const rows = dataPage.data;
+  if (!rows) return null;
 
   return (
     <div className="flex flex-col text-center">
       <div className="mb-5">
-        <div className="flex gap-3 justify-center">
-          <FaArrowLeft onClick={() => dataPage.prevPage?.()} />
-          <FaArrowRight onClick={() => dataPage.nextPage?.()} />
+        <div className="flex gap-3 justify-center select-none">
+          <FaArrowLeft
+            onClick={(e: MouseEvent) => changePage(e, "prev")}
+            className={`${dataPage.prevPage ? "cursor-pointer" : "opacity-50"}`}
+          />
+          <FaArrowRight
+            onClick={(e: MouseEvent) => changePage(e, "next")}
+            className={`${dataPage.nextPage ? "cursor-pointer" : "opacity-50"}`}
+          />
         </div>
-        <div className="max-w-full overflow-auto py-3">
+        <div
+          className={`max-w-full overflow-auto py-3 ${
+            dataPage.isLoading ? "opacity-50" : ""
+          }`}
+        >
           <table className="table-fixed text-left">
             <thead>
               <tr className="">
-                {Object.keys(dataPage.data[0]).map((key) => (
+                {Object.keys(rows[0]).map((key) => (
                   <th className="px-3" key={key}>
                     {key}
                   </th>
                 ))}
               </tr>
             </thead>
-            <tbody>
-              {dataPage.data.map((row, i) => (
+            <tbody className="h-[15rem]">
+              {rows.map((row, i) => (
                 <tr key={i}>
                   {Object.keys(row).map((key) => {
                     let value: string;
-                    if (typeof row[key] === 'object') {
+                    if (typeof row[key] === "object") {
                       value = JSON.stringify(row[key]);
-                    } else if (typeof row[key] === 'boolean') {
-                      value = row[key] ? 'true' : 'false';
+                    } else if (typeof row[key] === "boolean") {
+                      value = row[key] ? "true" : "false";
                     } else {
                       value = row[key];
                     }
@@ -77,7 +95,7 @@ export default function ManageData({
         <input
           className="border-2 rounded border-gray-400 px-2 text-center w-20 flex-auto"
           onChange={(e) => {
-            if (e.target.value === 'I am certain') setCanDelete(true);
+            if (e.target.value === "I am certain") setCanDelete(true);
             else setCanDelete(false);
           }}
         ></input>
